Move UI logic to view model
